Adds register and un-register to registration
This adds in some validation to a registration from the actions layer
and allows a registration to be completed and stored or returns
an error to be handled in the handler.
Thanks to @juanpaco for clarifying where best to validate
the data flowing into the system when using the HAQT pattern.

Todd Pickell <tapickell@gmail.com> @tapickell